requirements.txt
pillow==11.0.0
prompt_toolkit==3.0.48
wcwidth==0.2.13
